Return User schema on sign up and sign in
Both sign up and sign in are handled by google oauth2 system.
Once the user sign up or sign in, the backend verifies if the g_id (oauth2 id) is already in the database. If not, it is registered as a new user. On both cases the User schema is returned.